Added check for backend status
When the app starts, checkBackendStatus() pings the backend to check its availability.
If the response is not OK, a 3 seconds interval is set to retry the check. Should the backend become available, the check stops.

While the backend is not available, the app is not useable and a message is displayed.